De-/select multiple layers at once ; use my own fork of leaflet-panel-layers
This commit fixes #20

---

Because now we see for a group at its checkbox if all layers in it are
selected, we can collapse the group in this case.

---

The maintainer of stefanocudini/leaflet-panel-layers doesn't seem to be
active anymore.

Therefore I made my own fork of it, added new features and released them
on npm.

The existing fork syonfox/leaflet-panel-layers has breaking layout
changes and is therefore incompatible with this project.