ie11: iframe to the rescue (*)
In IE11 the only way to disable the referrer seems to be `<meta
name="referrer" content="never">`.

But I only want this for images and not globally for all resources. So I
wrap it inside of an `<iframe>` to separate it from the other content.

Change in behaviour:
- Before this change it was increasing the size of the popup after
fetching/loading the image, and then reloading the popup
(`marker._popup.update()`).
- Now it reserves the maximum possible space before and then decreases
the size after fetching, without doing a popup reload. Because with an
iframe with its own scope, the reload would result in the image being
fetched two times everytime a popup opens.

Though a general downside of the iframes compared to before, is that it
fetches the image multiple times when the same popup is opened again.
But that's what the browser cache is for.

This commit fixes issue #34.