collapse language menu (*)
If the device is not touchable it is based on :hover.
A mouse pointer is assumed and the menu expands/collapses with the mouse movement.

If the device is touchable (mobile) it is based on :focus.
The menu expands by tapping the current language.
It collapses, when tapping out of it or by choosing another language (this'd not be possible with :hover).

This commit adds to issue #11
This commit fixes issue #42