workaround: link to wiki images directly again
For anonymous users `File:` links do redirect to random article pages that use the files. Meaning that links to specific files are useless now. (Not
random, but the first article using it or sometimes the wiki main page. Which is in no way controllable.)
See https://community.fandom.com/f/p/3685233706663939045 for more details (they favor SEO over UX, which is a bullshit move...).

A `?file=` link will at least open the specific file in a dialog (except on IE -.-).
Luckily that works on any page, so I can open the image dialog on pages where the image isn't used (fallback: main page).

On IE the whole wiki is broken anyway since the FandomDesktop template migration, so this change shouldn't make the UX worse. Maybe it's even better
now, because the default redirect isn't always working that great (sometimes going to the wiki main page).